// resources/js/merchant/generated/src/models/CreateMerchantWithdrawalRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * User Wallet API
 * API for User Wallet application with order and transaction management
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: admin@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateMerchantWithdrawalRequest
 */
export interface CreateMerchantWithdrawalRequest {
    /**
     * 
     * @type {number}
     * @memberof CreateMerchantWithdrawalRequest
     */
    amount: number;
    /**
     * 
     * @type {string}
     * @memberof CreateMerchantWithdrawalRequest
     */
    description?: string;
}

/**
 * Check if a given object implements the CreateMerchantWithdrawalRequest interface.
 */
export function instanceOfCreateMerchantWithdrawalRequest(value: object): value is CreateMerchantWithdrawalRequest {
    if (!('amount' in value) || value['amount'] === undefined) return false;
    return true;
}

export function CreateMerchantWithdrawalRequestFromJSON(json: any): CreateMerchantWithdrawalRequest {
    return CreateMerchantWithdrawalRequestFromJSONTyped(json, false);
}

export function CreateMerchantWithdrawalRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateMerchantWithdrawalRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'amount': json['amount'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function CreateMerchantWithdrawalRequestToJSON(json: any): CreateMerchantWithdrawalRequest {
    return CreateMerchantWithdrawalRequestToJSONTyped(json, false);
}

export function CreateMerchantWithdrawalRequestToJSONTyped(value?: CreateMerchantWithdrawalRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'amount': value['amount'],
        'description': value['description'],
    };
}